fix VS warnings

+++ chips/_z80_opcodes.h
@@ -22,7 +22,7 @@ static uint8_t _z80_szp[256] = {
 /* set 16-bit address and 8-bit data in 64-bit pin mask */
 #define _SAD(addr,data) pins=(pins&~0xFFFFFF)|(((data)<<16)&0xFF0000ULL)|((addr)&0xFFFFULL)
 /* extract 8-bit data from 64-bit pin mask */
-#define _GD() ((pins&0xFF0000ULL)>>16)
+#define _GD() ((uint8_t)((pins&0xFF0000ULL)>>16))
 /* enable control pins */
 #define _ON(m) pins|=(m)
 /* disable control pins */
